base has been renamed to auth
