+++ pyproject.toml
@@ -51,6 +51,7 @@ target-version = ['py310']
 [tool.isort]
 profile = "black"
 line_length = 80
+skip_glob = "^(tests|splatnet3_scraper)"
 
 [tool.mypy]
 python_version = "3.10"
