+++ splatnet3_scraper/auth/exceptions.py
