+++ splatnet3_scraper/auth/graph_ql_queries.py
