--- splatnet3_scraper/base/tokens/__init__.py
@@ -1,2 +1,0 @@
-from splatnet3_scraper.base.tokens.nso import NSO
-from splatnet3_scraper.base.tokens.token_manager import Token, TokenManager
