+++ tests/auth/test_token_manager.py
@@ -7,13 +7,13 @@ import pytest
 import pytest_mock
 import requests
 
-from splatnet3_scraper.base.exceptions import (
+from splatnet3_scraper.auth.exceptions import (
     NintendoException,
     SplatnetException,
 )
-from splatnet3_scraper.base.graph_ql_queries import GraphQLQueries
-from splatnet3_scraper.base.tokens.nso import NSO
-from splatnet3_scraper.base.tokens.token_manager import Token, TokenManager
+from splatnet3_scraper.auth.graph_ql_queries import GraphQLQueries
+from splatnet3_scraper.auth.nso import NSO
+from splatnet3_scraper.auth.token_manager import Token, TokenManager
 from tests.mock import MockNSO
 
 
@@ -23,7 +23,7 @@ def mock_token(token_type):
 
 
 base_path = pathlib.Path(__file__).parent.parent / "fixtures" / "config_files"
-token_manager_path = "splatnet3_scraper.base.tokens.token_manager.TokenManager"
+token_manager_path = "splatnet3_scraper.auth.token_manager.TokenManager"
 
 
 class TestToken:
